# be/docs/swagger.yaml
basePath: /api
definitions:
...
  dto.CompanyResponse:
    properties:
      address:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      logo:
        type: string
      name:
        type: string
      npwp:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.CreatePositionRequest:
    properties:
      base_salary:
        type: number
      company_id:
        type: string
      department_id:
        type: string
      name:
        type: string
    required:
    - company_id
    - name
    type: object
  dto.CreateShiftRequest:
    properties:
      company_id:
        type: string
      end_time:
        type: string
      name:
        type: string
      start_time:
        type: string
    required:
    - company_id
    - end_time
    - name
    - start_time
    type: object
...
  dto.DepartmentResponse:
    properties:
      company:
        $ref: '#/definitions/dto.CompanyResponse'
      company_id:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.PositionResponse:
    properties:
      base_salary:
        type: number
      company:
        $ref: '#/definitions/dto.CompanyResponse'
      company_id:
        type: string
      created_at:
        type: string
      department:
        $ref: '#/definitions/dto.DepartmentResponse'
      department_id:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.ShiftResponse:
    properties:
      company:
        $ref: '#/definitions/dto.CompanyResponse'
      company_id:
        type: string
      created_at:
        type: string
      end_time:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      start_time:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdatePositionRequest:
    properties:
      base_salary:
        type: number
      department_id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
    type: object
  dto.UpdateShiftRequest:
    properties:
      end_time:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      start_time:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@hris.com
    name: API Support
  description: Human Resource Information System API with authentication and comprehensive
    HRIS modules
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: HRIS API
  version: "1.0"
paths:
...
  /positions:
    get:
      description: Retrieve all positions, optionally filtered by company
      parameters:
      - description: Filter by company ID
        in: query
        name: company_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Positions retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.PositionResponse'
                  type: array
              type: object
        "500":
          description: Failed to fetch positions
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get all positions
      tags:
      - Positions
    post:
      consumes:
      - application/json
      description: Create a new position
      parameters:
      - description: Position data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePositionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Position created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PositionResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create a position
      tags:
      - Positions
  /positions/{id}:
    delete:
      description: Delete a position by ID
      parameters:
      - description: Position ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Position deleted
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to delete
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete a position
      tags:
      - Positions
    get:
      description: Retrieve a position by its ID
      parameters:
      - description: Position ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Position retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PositionResponse'
              type: object
        "404":
          description: Position not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get position by ID
      tags:
      - Positions
    put:
      consumes:
      - application/json
      description: Update an existing position by ID
      parameters:
      - description: Position ID
        in: path
        name: id
        required: true
        type: string
      - description: Position data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdatePositionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Position updated
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.PositionResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update a position
      tags:
      - Positions
  /shifts:
    get:
      description: Retrieve all shifts, optionally filtered by company
      parameters:
      - description: Filter by company ID
        in: query
        name: company_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Shifts retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ShiftResponse'
                  type: array
              type: object
        "500":
          description: Failed to fetch shifts
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get all shifts
      tags:
      - Shifts
    post:
      consumes:
      - application/json
      description: Create a new work shift
      parameters:
      - description: Shift data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateShiftRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Shift created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ShiftResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Create a shift
      tags:
      - Shifts
  /shifts/{id}:
    delete:
      description: Delete a shift by ID
      parameters:
      - description: Shift ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Shift deleted
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to delete
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Delete a shift
      tags:
      - Shifts
    get:
      description: Retrieve a shift by its ID
      parameters:
      - description: Shift ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Shift retrieved
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ShiftResponse'
              type: object
        "404":
          description: Shift not found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Get shift by ID
      tags:
      - Shifts
    put:
      consumes:
      - application/json
      description: Update an existing shift by ID
      parameters:
      - description: Shift ID
        in: path
        name: id
        required: true
        type: string
      - description: Shift data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateShiftRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Shift updated
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ShiftResponse'
              type: object
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - Bearer: []
      summary: Update a shift
      tags:
      - Shifts
...
schemes:
- http
- https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"